Minor changes recommended/caught by gin66
The only real "big" thing was the change to isRunnig, which always
returned true, caused the need for startPreparedQueue, and a small
change in order to commandAddedToQueue (moving the write pointer
increment outside of the conditional of start)

I also made the volatile rp and wp for SAM only.  Again, it should be
just a hint, and should not remove any optimizations beyond variable
caching, but gin66 is sure its unnecessary.  Its the kind of thing that
requires some longevity tests though, so I just ifdef'd it for the Due.
Unless the compiler for the AVR/ESP is written incorrectly, this should
not be able to impact performance on these architecturees unless it was
incorrectly caching the variable!

I've restested (short term) that these changes are still functional.