Fix analogWrite being clobbered by the PWM handler
Also stopped setting the CLKB register, because I'm using MCK/4 which is always provided to all PWM channels regardless, so I don't need to stick it into ClockB!  Changing the PWM_CLK register while a channel was active sometimes caused bad behavior even if the clock that channel was using was not changed.

I edited these changes in place, they need to be tested.  I'm hoping @sigmaeo can test with a clean copy.  I have numerous other small changes that do not affect the function of the library I'd like to keep.  But, I'll save and restore it if necessary.  I'd also like @sigmaeo to test and close the issue if the fix works :)