Fix attempt for https://github.com/gin66/FastAccelStepper/issues/250
On quick direction changes a direction toggle could be missed.
My understanding is that this code branch "restarts" the motion right at a stop event. But the direction change would not be applied either doing a few steps in the wrong direction, or if a large move was initiated at this time a much more impacting movement in the wrong direction.

If it's motion restart it feels correct to toggle direction before that ultimate step, but i'm not 100% sure.